% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCellNumbs.R
\name{calcCellNumb}
\alias{calcCellNumb}
\title{Calculate cell numbers}
\usage{
calcCellNumb(N0 = 1, d = 24, t)
}
\arguments{
\item{N0}{Initial number of cells. Default = 1}

\item{d}{Doubling time of the cell culture. Default = 24 hours}

\item{t}{Time in hours}
}
\value{
The number of cells after \code{t} hours
}
\description{
Calculates the number of cells after \code{t} hours, by using the 
doubling time as a measure of mitosis and death of cells in the culture
}
\details{
The function calculates the number of cells by using the equation 
\eqn{N = N0 * 2^(t/d)}
}
\examples{
calcCellNumber(1,24,24) = 2
}
