% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIntersects.R
\name{calcIntersects}
\alias{calcIntersects}
\title{Calculate mutations and their probabilities}
\usage{
calcIntersects(r, hline)
}
\arguments{
\item{r}{Data.frame with the results from \code{\link{probDriverMut}}}

\item{hline}{Numeric indicating the value on the y-axis to intersect with}
}
\value{
Numeric vector of intersections with \code{hline} for different cell types and genes
}
\description{
Calculate the number of mutations and the probabilities on this number of mutations per cells.
}
\examples{
invitro_res = probDriverMut(cell_type = cell_types, gene = names(driver_counts), stat_type = "mean")
hline = 1
calcIntersect(r = invitro_res, hline = hline)

}
