% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProbabilities.R
\name{calcProbabilities}
\alias{calcProbabilities}
\title{Calculate probability of obtaining driver gene mutation}
\usage{
calcProbabilities(N, N0 = 1, mu = NULL, d = NULL, mt = NULL,
  dep = NULL, npos = NULL, perc = 0.015, psnv = NULL,
  count = NULL, invivo = F)
}
\arguments{
\item{N}{Integer giving the generation number}

\item{N0}{Integer giving the initial number of cells}

\item{mu}{Integer giving the mutation rate of the cell type}

\item{d}{Integer giving the doubling time of the cell type}

\item{mt}{Integer giving the mitosis rate of the cell type}

\item{dep}{Integer giving the depletion of mutations in the Coding Sequence}

\item{npos}{Integer giving the length of the Coding Sequence}

\item{perc}{Integer giving the percentage of coding sequence in whole genome}

\item{psnv}{Table with the mutation spectrum from the cell type}

\item{count}{Table with the number of driver gene mutations specified per mutation type}
}
\value{
Number of mutations accumulated after N generations
}
\description{
Calculate probability of obtaining driver gene mutation
}
