% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDriverSNVs.R
\name{countDriverSNVs}
\alias{countDriverSNVs}
\title{Counting driver activating mutations}
\usage{
countDriverSNVs(drivers, per_gene = F, tri = F)
}
\arguments{
\item{drivers}{Dataframe with 2 columns. First column must be gene ID and second the mutations}

\item{per_gene}{Indicating if driver mutations are counted per gene, default is FALSE}
}
\value{
Table of counts of the 6 mutation types 'C>A', 'C>G, 'C>T, 'T>A', 'T>C', 'T>G'
}
\description{
Counts the number of driver activating mutations in a gene or in total
}
\examples{
drivers = data.frame(gene=c('ABL1','ABL1'),gdna=c('G>A','A>G'), stringsAsFactors=FALSE)
countDriverSNVs(drivers)
}
