% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDriverSNVs.R
\name{getDriverSNVs}
\alias{getDriverSNVs}
\title{Calculate mutations and their probabilities}
\usage{
getDriverSNVs(table1, table2)
}
\arguments{
\item{cell_type}{Character indicating the cell type}

\item{gene}{Character indicating which gene to analyse}

\item{stat_type}{Character indicating the type of statistic to calculate, must be "mean",
"mean-std" or "mean+std"}
}
\value{
Data.frame with the number of cells, the number of mutations, the probabilities of 
the mutations and the input variables
}
\description{
Calculate the number of mutations and the probabilities on this number of mutations per cells.
}
\examples{
probDriverMut(cell_type = "iPS", gene = "all", stat_type = "mean")

}
