% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invivoYears.R
\name{invivoYears}
\alias{invivoYears}
\title{Plot in vivo results}
\usage{
invivoYears(r, gene = "all", dep_invitro, CDS_length)
}
\arguments{
\item{r}{Dataframe with results from \code{\link{probDriverMut}}}

\item{gene}{Character string to indicate which gene must be analyzed}

\item{dep_invitro}{List with enrichment/depletion test results of invitro experiments}

\item{CDS_length}{Numeric with length of Coding Sequence of 'in vivo' cell type}
}
\value{
Extended version of results from \code{\link{probDriverMut}}, with
in column year the amount of years in vivo
}
\description{
Function to plot \(a subset of\) the number of in vitro grown SI ASC, with the
probability on oncogenic mutation, compared to number of years of in vivo 
mutation accumulation of SI ASC.
}
\examples{
r = probDriverMut(cell_type = cell_types, gene = "all")

invivo_plot = plotInvivoResults(r, gene = "all", dep_invitro = CDS_dep, 
                                CDS_length = CDS_length[["SI"]])

}
