% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotInvivoResults.R
\name{plotInvivoResults}
\alias{plotInvivoResults}
\title{Plot in vivo results}
\usage{
plotInvivoResults(r, cell_type = "SI", gene = "all", color = "black",
  xlab = "Probability", ylab = "Number of cells", title = "",
  pdf_out)
}
\arguments{
\item{r}{Dataframe with results from \code{\link{probDriverMut}}}

\item{gene}{Character string to indicate which gene must be plotted}

\item{color}{Character string for color of line in figure}

\item{xlab}{Character string with label for x-axis. Default is\cr
"Number of cells"}

\item{ylab}{Character string with label for y-axis. Default is\cr
"Number of SNVs in oncogenes"}

\item{title}{Character string with title of figure. Default is no title}

\item{pdf_out}{Character with file name where to save the figure. If not given,
figure will not be saved, only returned from function}
}
\value{
Plot of probablity on oncogenic mutations against cell numbers and years
of in vivo mutation accumulation
}
\description{
Function to plot \(a subset of\) the number of in vitro grown SI ASC, with the
probability on oncogenic mutation, compared to number of years of in vivo 
mutation accumulation of SI ASC.
}
\examples{
r = data.frame(cells = calcCellNumb, muts = calcMutEvents, probs = probDriverMut)

invivo_plot = plotInvivoResults(r, gene = "all", color = "red")

}
