% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMutationsPerCell.R
\name{plotMutationsPerCell}
\alias{plotMutationsPerCell}
\title{Plot results per cell}
\usage{
plotMutationsPerCell(r, xmax, ymax, logscale = c("x", "y"), vline = F,
  hline = F, xlab = "Number of cells",
  ylab = "Number of SNVs in oncogenes", title = "", pdf_out)
}
\arguments{
\item{r}{Dataframe with results from \code{\link{probDriverMut}}}

\item{xmax}{Numeric with x-axis maximum}

\item{ymax}{Numeric with y-axis maximum}

\item{logscale}{(Optional) Character vector indicating which axes must be set to 
logscale. Default is \code{c("x","y")}}

\item{vline}{Numeric vector with vertical lines}

\item{hline}{Numeric vector with horizontal lines}

\item{xlab}{Character string with label for x-axis. Default is\cr
"Number of cells"}

\item{ylab}{Character string with label for y-axis. Default is\cr
"Number of SNVs in oncogenes"}

\item{title}{Character string with title of figure. Default is\cr
"Mutation accumulation of driver gene mutations"}

\item{pdf_out}{Character with file name where to save the figure. If not given,
figure will not be saved, only returned from function}
}
\value{
Plot of mutational events against cell numbers.
}
\description{
Function to plot \(a subset of\) the results from \code{\link{calcCellNumb}}, \code{\link{calcMutEvents}} and 
\code{\link{probDriverMut}}. Unit is cells
}
\examples{
r = data.frame(cells = calcCellNumb, muts = calcMutEvents, probs = probDriverMut)

mutation_plot = plotMutationsPerCell(r, xmax = 1e13, ymax = 1e4, hline = 1, vline = 1e6)
}
