% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probDriverMut.R
\name{probDriverMut}
\alias{probDriverMut}
\title{Calculate mutations and their probabilities}
\usage{
probDriverMut(cell_types, genes, stat_type = "mean", verbose = F)
}
\arguments{
\item{cell_types}{Character vector indicating the cell types}

\item{genes}{Character vector indicating which genes to analyse}

\item{verbose}{Boolean indicating verbose mode}

\item{stat_types}{Character vector indicating the types of statistic to calculate:
"mean", "mean-std" and/or "mean+std"}
}
\value{
Data.frame with the number of cells, the number of mutations, the probabilities of 
the mutations and the input variables
}
\description{
Calculate the number of mutations and the probabilities on this number of mutations per cells.
}
\examples{
probDriverMut(cell_type = "iPS", gene = "all", stat_type = "mean")

}
