% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{processData}
\alias{processData}
\title{Process incoming data}
\usage{
processData(mutation_rate_file, process_drivers = T,
  oncogenic_mutations_file, gene_MoA_file, prob_per_gene = F,
  trinucleotide = F, CDS_dep_invitro_file, CDS_dep_invivo_file,
  sample_celltypes_file, invivo_tissue, mut_spec_invitro_file,
  mut_spec_invivo_file)
}
\arguments{
\item{mutation_rate_file}{Character indicating the file containing column 'Cell_type' and
'Number_of_mutations_per_cell_cycle'}

\item{oncogenic_mutations_file}{Character indicating the file containing a oncogene driver
catalog with columns 'gene' and 'gdna' containing the variants}

\item{gene_MoA_file}{Character indicating the file containing columns 'gene' and 'gene_MoA'.
The column 'gene_MoA' contains either 'Act', 'ambiguous' or 'LoF' as mode of action of gene}

\item{prob_per_gene}{Boolean indicating if probability must be calculated per gene or for 
all genes together}

\item{trinucleotide}{Boolean indicating if trinucleotide context of drivers must be computed}

\item{CDS_dep_invitro_file}{Character indicating the file with the enrichment/depletion test
results from the mutationalPatterns package for invitro data}

\item{CDS_dep_invivo_file}{Character indicating the file with the enrichment/depletion test
results from the mutationalPatterns package for invivo data}

\item{invivo_tissue}{Character indicating the invivo tissue}

\item{mut_spec_invitro_file}{Character indicating the file with the invitro mutational spectrum}

\item{sample_cell_type_file}{Character indicating the file with the 2 colums 'Sample' and 'Type'}

\item{mut_spec_invitro_file}{Character indicating the file with the invivo mutational spectrum}
}
\value{
All necessary data for processing downstream will be imported as global variables
}
\description{
Read the data in the given files and process it to data.frames or vector needed by 
downstream functions. Processing of drivers is optional, if 'process_drivers' is set to False
a list of driver mutations must be given later on.
}
